refactor:sidebar profile ui 위치 entity로 변경
