+++ src/entities/user/ui/Profile.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { useGetProfileQuery } from '../model/useProfileQuery';
+import { useGetProfileQuery } from '../../../features/sidebar/profile/model/useProfileQuery';
 
 export default function Profile() {
   const { isLoading, data, error } = useGetProfileQuery();
